diyblog/requirements.txt
asgiref==3.2.7
boto3==1.13.6
botocore==1.16.6
certifi==2020.6.20
Collectfast==2.2.0
dj-database-url==0.5.0
Django==3.0.7
django-comments-xtd==2.4.0
django-contrib-comments==1.9.2
django-js-asset==1.2.2
django-secure==1.0.1
django-storages==1.9.1
django-summernote==0.8.11.6
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.9.4
Pillow==7.1.2
psycopg2-binary==2.8.5
python-dateutil==2.8.0
pytz==2020.1
s3transfer==0.3.3
six==1.12.0
sqlparse==0.3.1
typing-extensions==3.7.4.2
urllib3==1.25.3
whitenoise==5.0.1
django-ckeditor==5.9.0
